v0.3-MCP-Server/requirements.txt
mcp[cli]>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0